// src/app/components/phone/types.ts
export type AppScreen =
  | "tictactoe"
  | "avatarid"
  | "folder"
  | "space"
  | "yeiza"
  | "calculator"
  | "clock"
  | "settings"
  | "camera"
  | "music"
  | "notes"
  | "messages"
  | "dialer"
  | "appstore"
  | "chrome"
  | "mibrowser"
  | "gmail"
  | "youtube";

export type Screen = "home" | "medium" | AppScreen;